ctrlport: make gr-digital not dependent on ctrlport.
We don't want components to require ctrlport, but if it is configured, we want them to use it. This only adds the libraries and include directories when it's available.

Note that digital_pfb_clock_sync_ccf will fail right now if gr-ctrlport is disabled since we aren't checking internally for it's existence.